+++ src/frontend/app/shared-tdoc/components/tdoc-linked-playlists/tdoc-linked-playlists.component.ts
@@ -2,7 +2,6 @@ import {ChangeDetectionStrategy, ChangeDetectorRef, Component} from '@angular/co
 import {TourDocRecord} from '../../../../shared/tdoc-commons/model/records/tdoc-record';
 import {DomSanitizer} from '@angular/platform-browser';
 import {CommonRoutingService} from '@dps/mycms-frontend-commons/dist/angular-commons/services/common-routing.service';
-import {CommonDocRoutingService} from '@dps/mycms-frontend-commons/dist/frontend-cdoc-commons/services/cdoc-routing.service';
 import {TourDocLinkedPlaylistRecord} from '../../../../shared/tdoc-commons/model/records/tdoclinkedplaylist-record';
 import {GenericAppService} from '@dps/mycms-commons/dist/commons/services/generic-app.service';
 import {TourDocActionTagService} from '../../services/tdoc-actiontag.service';
@@ -10,7 +9,10 @@ import {ToastrService} from 'ngx-toastr';
 import {TourDocSearchForm} from '../../../../shared/tdoc-commons/model/forms/tdoc-searchform';
 import {TourDocSearchResult} from '../../../../shared/tdoc-commons/model/container/tdoc-searchresult';
 import {TourDocDataService} from '../../../../shared/tdoc-commons/services/tdoc-data.service';
-import {CommonDocLinkedPlaylistsComponent} from '@dps/mycms-frontend-commons/dist/frontend-cdoc-commons/components/cdoc-linked-playlists/cdoc-linked-playlists.component';
+import {
+    CommonDocLinkedPlaylistsComponent
+} from '@dps/mycms-frontend-commons/dist/frontend-cdoc-commons/components/cdoc-linked-playlists/cdoc-linked-playlists.component';
+import {TourDocRoutingService} from '../../../../shared/tdoc-commons/services/tdoc-routing.service';
 
 @Component({
     selector: 'app-tdoc-linked-playlists',
@@ -22,10 +24,10 @@ export class TourDocLinkedPlaylistsComponent extends CommonDocLinkedPlaylistsCom
     TourDocSearchResult, TourDocDataService, TourDocActionTagService, TourDocLinkedPlaylistRecord> {
 
     constructor(sanitizer: DomSanitizer, commonRoutingService: CommonRoutingService,
-                cdocRoutingService: CommonDocRoutingService, appService: GenericAppService,
+                protected tdocRoutingService: TourDocRoutingService, appService: GenericAppService,
                 actionTagService: TourDocActionTagService, toastr: ToastrService,
                 cdocDataService: TourDocDataService, cd: ChangeDetectorRef) {
-        super(sanitizer, commonRoutingService, cdocRoutingService, appService, actionTagService, toastr, cdocDataService, cd);
+        super(sanitizer, commonRoutingService, tdocRoutingService, appService, actionTagService, toastr, cdocDataService, cd);
     }
 
     protected updateData(): void {
@@ -40,4 +42,13 @@ export class TourDocLinkedPlaylistsComponent extends CommonDocLinkedPlaylistsCom
         this.linkedPlaylists = this.record['tdoclinkedplaylists'];
     }
 
+    public submitShow(event, linkedPlaylist: TourDocLinkedPlaylistRecord): boolean {
+        this.tdocRoutingService.navigateToShow(this.generateRecord(linkedPlaylist), '');
+        return false;
+    }
+
+    private generateRecord(linkedPlaylist: TourDocLinkedPlaylistRecord): TourDocRecord {
+        return this.cdocDataService.newRecord({id: 'PLAYLIST_' + linkedPlaylist.refId,
+            name: linkedPlaylist.name, type: 'PLAYLIST'});
+    }
 }