+++ src/shared/tdoc-commons/model/datamode.enum.ts
@@ -1,0 +1,3 @@
+export enum DataMode {
+    BACKEND, STATIC
+}